rework, add presentation, add sub-type enums
