+++ src/dash-table/syntax-tree/lexeme/block.ts
@@ -17,7 +17,7 @@ export const blockOpen: IUnboundedLexeme = {
     },
     type: LexemeType.BlockOpen,
     nesting: 1,
-    present: '()',
+    present: () => '()',
     priority: 1,
     regexp: /^\(/,
     syntaxer: (lexs: any[]) => {
