--- src/dash-table/syntax-tree/lexeme/or.ts
@@ -1,25 +1,0 @@
-import Logger from 'core/Logger';
-import { IUnboundedLexeme, LexemeType } from 'core/syntax-tree/lexicon';
-
-const or: IUnboundedLexeme = {
-    evaluate: (target, tree) => {
-        Logger.trace('evaluate -> ||', target, tree);
-
-        const t = tree as any;
-
-        return t.left.lexeme.evaluate(target, t.left) ||
-            t.right.lexeme.evaluate(target, t.right);
-    },
-    type: LexemeType.LogicalOperator,
-    present: '||',
-    priority: 3,
-    regexp: /^(or\s|\|\|)/i,
-    syntaxer: (lexs: any[], pivot: any, pivotIndex: number) => {
-        return Object.assign({
-            left: lexs.slice(0, pivotIndex),
-            right: lexs.slice(pivotIndex + 1)
-        }, pivot);
-    }
-};
-
-export default or;
